move rollup configs to `configs/rollup` (#1462)
