+++ configs/rollup/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "configs",
+	"name": "@configs/rollup",
 	"version": "0.1.0",
 	"private": true,
 	"type": "module",
