+++ libs/@guardian/ab-core/package.json
@@ -46,7 +46,7 @@
 		"build": {
 			"command": "rollup -c",
 			"files": [
-				"../../../configs/rollup.config.js",
+				"../../../configs/rollup/rollup.config.js",
 				"./rollup.config.js",
 				"src/**",
 				"package.json",
