+++ libs/@guardian/eslint-plugin-source-foundations/package.json
@@ -57,7 +57,7 @@
 		"build": {
 			"command": "rollup -c",
 			"files": [
-				"../../../configs/rollup.config.js",
+				"../../../configs/rollup/rollup.config.js",
 				"./rollup.config.js",
 				"src/**",
 				"package.json",
