+++ libs/@guardian/eslint-plugin-source-react-components/package.json
@@ -60,7 +60,7 @@
 		"build": {
 			"command": "rollup -c",
 			"files": [
-				"../../../configs/rollup.config.js",
+				"../../../configs/rollup/rollup.config.js",
 				"./rollup.config.js",
 				"src/**",
 				"package.json",
