+++ libs/@guardian/identity-auth-frontend/package.json
@@ -54,7 +54,7 @@
 				"../identity-auth:build"
 			],
 			"files": [
-				"../../../configs/rollup.config.js",
+				"../../../configs/rollup/rollup.config.js",
 				"./rollup.config.js",
 				"src/**",
 				"package.json",
