+++ libs/@guardian/identity-auth/package.json
@@ -52,7 +52,7 @@
 		"build": {
 			"command": "rollup -c",
 			"files": [
-				"../../../configs/rollup.config.js",
+				"../../../configs/rollup/rollup.config.js",
 				"./rollup.config.js",
 				"src/**",
 				"package.json",
