+++ libs/@guardian/newsletter-types/rollup.config.js
@@ -1,3 +1,3 @@
-import baseConfigs from '../../../configs/rollup.config.js';
+import baseConfigs from '../../../configs/rollup/rollup.config.js';
 
 export default baseConfigs()[2];
