+++ libs/@guardian/source-foundations/rollup.config.js
@@ -1,1 +1,1 @@
-export * as default from '../../../configs/rollup.config.js';
+export * as default from '../../../configs/rollup/rollup.config.js';
