+++ libs/@guardian/source-react-components-development-kitchen/package.json
@@ -67,7 +67,7 @@
 				"_deps"
 			],
 			"files": [
-				"../../../configs/rollup.config.js",
+				"../../../configs/rollup/rollup.config.js",
 				"./rollup.config.js",
 				"src/**",
 				"package.json",
