+++ libs/@guardian/source-react-components/package.json
@@ -78,7 +78,7 @@
 				"_deps"
 			],
 			"files": [
-				"../../../configs/rollup.config.js",
+				"../../../configs/rollup/rollup.config.js",
 				"./rollup.config.js",
 				"src/**",
 				"package.json",
