+++ libs/@guardian/source/package.json
@@ -91,7 +91,7 @@
 				"_deps"
 			],
 			"files": [
-				"../../../configs/rollup.config.js",
+				"../../../configs/rollup/rollup.config.js",
 				"./rollup.config.js",
 				"src/**",
 				"package.json",
